Implemented first dozer mappings
- implemented first dozer mappings
- remark! some code changes prevent building at the moment

+++ brusselnieuws-rss/src/main/java/be/seriousbusiness/brusselnieuws/rss/datasource/dto/MediumDTO.java
@@ -7,7 +7,7 @@ public interface MediumDTO {
 	void setLink(final String link);
 	MediumTypeDTO getMediumTypeDTO();
 	void setMediumTypeDTO(final MediumTypeDTO mediumTypeDTO);
-	double getSize();
-	void setSize(final double size);
+	long getSize();
+	void setSize(final long size);
 
 }
